created playlist features
