+++ src/main/java/MusicApp/controller/UserController.java
@@ -40,9 +40,9 @@ public class UserController {
         return userService.deleteById(userId);
     }
 
-    //addSong is creating a song.
-    @PutMapping("/user/{username}/{songId}")
-    public User addSong(@PathVariable String username, @PathVariable int songId){
-        return userService.addSong(username, songId);
-    }
+    //addSong is attaching a song to a user
+//    @PutMapping("/user/{username}/{songId}")
+//    public User addSong(@PathVariable String username, @PathVariable int songId){
+//        return userService.addSong(username, songId);
+//    }
 }
