+++ src/main/java/MusicApp/models/User.java
@@ -72,18 +72,22 @@ public class User {
 
     public void setUserRole(UserRole userRole) { this.userRole = userRole;}
 
-    @ManyToMany(fetch = FetchType.LAZY, cascade = {CascadeType.DETACH, CascadeType.MERGE, CascadeType.REFRESH})
-    @JoinTable(name = "user_song", joinColumns = {@JoinColumn(name = "user_id")}, inverseJoinColumns = @JoinColumn(name = "song_id"))
-    private List<Song> songs;
 
-    public List<Song> getSong(){ return songs; }
+    //Commenting this out to transfer this functionality to the userprofile
+//    @ManyToMany(fetch = FetchType.LAZY, cascade = {CascadeType.DETACH, CascadeType.MERGE, CascadeType.REFRESH})
+//    @JoinTable(name = "user_song", joinColumns = {@JoinColumn(name = "user_id")}, inverseJoinColumns = @JoinColumn(name = "song_id"))
+//    private List<Song> songs;
 
-    public void setSong(List<Song> song) { this.songs = song; }
+//    public List<Song> getSong(){ return songs; }
+//
+//    public void setSong(List<Song> song) { this.songs = song; }
 
-    public List<Song> addSong(Song song){
-        if(songs == null)
-            songs = new ArrayList<>();
-            songs.add(song);
-        return songs;
-    }
+//    public List<Song> addSong(Song song){
+//        if(songs == null)
+//            songs = new ArrayList<>();
+//            songs.add(song);
+//        return songs;
+//    }
+
+
 }
