+++ src/main/java/MusicApp/repositories/UserProfileRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 public interface UserProfileRepository extends CrudRepository<UserProfile, Long> {
 
-    @Query("FROM UserProfile up LEFT JOIN User u ON u.username = ?1 AND up.id = u.userProfile.id")
+    @Query("FROM UserProfile up JOIN User u ON u.username = ?1 AND up.id = u.userProfile.id")
     public UserProfile findProfileByUsername(String username);
 }
