+++ src/main/java/MusicApp/service/UserProfileService.java
@@ -5,4 +5,5 @@ import MusicApp.models.UserProfile;
 public interface UserProfileService {
     public UserProfile createUserProfile(String username, UserProfile newProfile);
     public UserProfile getUserProfile(String username);
+    public UserProfile addSong(String username, int songId);
 }
