+++ src/main/java/MusicApp/service/UserProfileServiceImpl.java
@@ -1,8 +1,11 @@
 package MusicApp.service;
 
+import MusicApp.models.Song;
 import MusicApp.models.User;
 import MusicApp.models.UserProfile;
+import MusicApp.repositories.SongRepository;
 import MusicApp.repositories.UserProfileRepository;
+import MusicApp.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +17,12 @@ public class UserProfileServiceImpl implements UserProfileService {
 
     @Autowired
     UserService userService;
+
+    @Autowired
+    UserRepository userRepository;
+
+    @Autowired
+    SongRepository songRepository;
 
     @Override
     public UserProfile createUserProfile(String username, UserProfile newProfile){
@@ -28,5 +37,16 @@ public class UserProfileServiceImpl implements UserProfileService {
     @Override
     public UserProfile getUserProfile(String username){
         return userProfileRepository.findProfileByUsername(username);
+    }
+
+
+    // This came from UserServiceImpl
+    @Override
+    public UserProfile addSong(String username, int songId) {
+        Song song = songRepository.findById(songId).get();
+        UserProfile userprofile = getUserProfile(username);
+        userprofile.addSong(song);
+
+        return userProfileRepository.save(userprofile);
     }
 }
